CloseAndStartAnotherPlugin rewritten using MacroSendMacroText

+++ nf3x/src/far/PanelInfoWrap.h
@@ -78,7 +78,7 @@ public:
 		);
 	}
 
-	void UpdateAndRedraw(bool bActivePanel, int nCurrentItem = 0, int nTopPanelItem = 0) {
+	void UpdateAndRedraw(bool bActivePanel, size_t nCurrentItem = 0, size_t nTopPanelItem = 0) {
 		PanelRedrawInfo pri; 
 		pri.CurrentItem = nCurrentItem;
 		pri.TopPanelItem = nTopPanelItem;
